requirements.txt
opencv-python==4.8.1.78
pyzbar==0.1.9 
flask
opencv-python==4.8.1.78
pyzbar==0.1.9
qrcode
supabase
gunicorn
python-dotenv
google-auth
google-auth-oauthlib
google-auth-httplib2
requests